refactor(AppLayout): move file to adhere to conventions.
